requirements.txt
PyQt5==5.15.9
pytsk3==20211028
diffusers==0.21.2
transformers==4.33.2
accelerate==0.23.0
torch==2.1.0
Pillow==10.1.0
scipy==1.11.2
scikit-learn==1.3.0
